Expand GitHub Actions
because: @dogancanbakir requested in a PR on the
prior commit that the naming conventions align
with other projectdiscovery projects. It seems that
adding some of the standard suite of workflows will
aid in quicker development and consistency across
projects.

this commit: Adds additional workflows in a similar
manner to the projectdiscovery/httpx project.